# docs/api.yml
openapi: 3.0.3
info:
  title: ESN Restful API Documentation
  version: '1.0'
  contact: {}
  description: |-
    ## ESN API - SA4

    Some simple guidelines for our RESTful apis

    1. all properties must use`camel_case`.
    2. use plural nouns as endpoint names
    3. no abbreviations, use full-name as much as possible to reduce confusion
servers:
  - url: 'https://esn-f20sa4.herokuapp.com/api'
  - url: 'http://localhost:6003/api'
paths:
...
  /users:
    get:
      summary: list user
      tags:
        - User
      parameters:
        - in: query
          name: username
          schema:
            type: string
          required: false
        - in: query
          name: status
          schema:
            type: string
            enum: [ok, help, emergency]
          required: false
        - in: query
          name: longitude
          schema:
            type: number
            format: float
          required: false
        - in: query
          name: latitude
          schema:
            type: number
            format: float
          required: false
        - in: query
          name: near
          schema:
            type: number
            format: float
          required: false
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
          required: false
        - in: query
          name: current
          schema:
            type: integer
            default: 0
          required: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginationResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
        '401':
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /users/{username}:
    patch:
      summary: Update specific user by username
      tags:
        - User
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
      requestBody:
        description: Update user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      summary: Get the status of the user
      tags:
        - User
      parameters:
        - in: path
          name: username
          schema:
            type: string
          required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Not Authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
        online:
          type: boolean
        status:
          type: string
          enum: [Ok, Help, Emergency]
        status_last_updated_at:
          type: string
          format: date
        qualification:
          type: array
          items:
            type: string
            enum:
              [
                'first aid responder',
                'healthcare worker',
                'machinery operator',
                'rescuer',
                'technician',
                'firefighter',
                'law enforcement officer',
              ]
        coordinates:
          type: object
          properties:
            longitude:
              type: number
              format: float
            latitude:
              type: number
              format: float
      required:
        - username
        - online
...
    PaginationResponse:
      type: object
      properties:
        total:
          type: integer
        current:
          type: integer
        limit:
          type: integer
        data:
          type: array
          items:
            type: object
      required:
        - total
        - current
        - limit
        - data
...
    Error:
      type: object
      properties:
        error:
          description: Type of the error
          type: string
          enum:
            - ValidationError
            - AuthenticationError
            - AuthorizationError
            - NotFoundError
        message:
          description: Displayable message for the error
          type: string
      required:
        - error
        - message
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
security:
  - BearerAuth: []
  - cookieAuth: []